pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "servertools"
version = '2.0.0'
description = "A package for routines performed by my home automation server"
authors = ["bobrock <bobrock@tuta.io>"]
license = "MIT"
readme = 'README.md'
repository = 'https://github.com/barretobrock/servertools'
packages = [
    { include = 'servertools' },
]
include = ["CHANGELOG.md"]

[tool.poetry.dependencies]
python = "^3.10"
kavalkilu = {git = "https://github.com/barretobrock/kavalkilu.git"}
slacktools = {git = "https://github.com/barretobrock/slacktools.git"}
reolink_api = {git = "https://github.com/barretobrock/reolink-python-api.git"}
pukr = { git = "https://github.com/barretobrock/pukr.git", branch = "main" }
Flask = "2.1.1"
Pillow = "9.1.0"
amcrest = "1.9.7"
beautifulsoup4 = "^4.11.1"
grafana-api = "^1.0.3"
imutils = "0.5.4"
lxml = "4.8.0"
markovify = "0.9.4"
meteocalc = "1.1.0"
moviepy = "1.0.3"
nltk = "^3.7"
numpy = "^1.22.3"
opencv-python = "4.5.5.64"
openwrt-luci-rpc = "1.1.11"
pandas = "^1.4.2"
phue = "1.1"
pygsheets = "^2.0.5"
pylunar = "0.6.0"
pyowm = "3.3.0"
python-dateutil = "2.8.2"
python-yr = "1.4.7.post2"
requests = "^2.27.1"
roku = "4.1.0"
selenium = "4.1.3"
speedtest-cli = "2.1.3"
sqlalchemy = "1.4.32"
yahoofinancials = "1.6"

[tool.poetry.dev-dependencies]
pre-commit = "^2.20.0"
pytest = "^7.1.1"
pytest-cov = "^3.0.0"
flake8 = "^4.0.1"
tox = "^3.24.5"

[tool.isort]
profile = 'black'
force_grid_wrap = 2
force_sort_within_sections = true
skip_gitignore = true
skip_glob = ['*_pb2.py', '*_pb2_grpc.py']

[tool.poetry.extras]
test = ["pytest"]
